pyproject.toml
[project]
name = "claude-whisper"
version = "0.1.2"
description = "Voice-controlled AI assistant interface for Claude Code using Apple MLX Whisper"
readme = "README.md"
license = "BSD-3-Clause"
authors = [
    { name = "Ashton Sidhu", email = "sidhuashton@gmail.com" }
]
keywords = ["claude", "whisper", "voice", "speech", "ai", "assistant", "mlx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.18",
    "desktop-notifier>=6.2.0",
    "loguru>=0.7.3",
    "mlx-whisper>=0.4.3",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pynput>=1.8.1",
]

[project.urls]
Homepage = "https://github.com/Ashton-Sidhu/claude-whisper"
Repository = "https://github.com/Ashton-Sidhu/claude-whisper"
Issues = "https://github.com/Ashton-Sidhu/claude-whisper/issues"


[project.scripts]
claude-whisper = "claude_whisper:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
claude-whisper = { workspace = true}

[dependency-groups]
dev = [
    "claude-whisper",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.6",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["claude_whisper"]
